feat: adjust responses for relation entities
